Enhancement: update heading--l text size
Resolves issues that occur as `heading--l` and `heading--l-alit` were both 24px

For more background see: https://gitlab.ebi.ac.uk/emblorg/backlog/issues/215#note_45386

We could also resolve some of these issues by manually setting the `.vf-content h2` size, but that would be a bit of a hack and not get updates from vf-design-tokens.

This will have some knockon effects to other patterns, sometimes it may be a "change for the better" other times we may need to update a pattern to use `--l-alt`

The Sketch paragraph styles may also need to be updated.