Chore: set percy to run only on develop
This limits the percy command to only run on commits to develop and compares against master.

We might also want to disable the github integration to stop comparissons on PRs, which has chewed through our alotment of screenshots.

This also set it only do mobile/desktop for a 33% savings :D

But it doesn't really matter until August as we've already gone through our quota!