Guidance on how (and when) to start using vf-core #391, #339
In this commit we make it clear(er) that the role of the Fractal part of vf-core is to be the component library.

It also works to cleanup the docs in general leverage the vf-welcome site, which itself is a work in progress.

In all, this narrows the role of docs within vf-core to be directly about the component library and lets vf-welcome be an umbrella site to support both vf-core and vf-eleventy, and likely other sites as well.