bug: vf-build-asset-copy-misses-js
This resolves a bug that was exposed by #509, where `gulp vf-fractal:build` was responsible for asset copying.

That was the wrong spot to do so an as 509 used `parallel` for `vf-fractal:build`, it was resulting the JS assets being missed from deployment.

Note that https://dev.assets.emblstatic.net/vf/develop/scripts/scripts.js is currently a 403 (missing and denied)