Feature: Drop vf-frctl-theme
I think this is one that's either a conceptual snapshot or something we just do.

As mentioned in #559, with vf-welcome in better shape. We no longer need to rely on vf-core for onboarding into the Visual Framework ecosystem.

That means we can drop the custom vf-frctl-theme; a quick go at pros/cons:

Pro:

- One less thing to maintain
- The Fractal templating was different that all other templates
- Allows the vf-core to send of visual clues that it's for low-level devs

Downsides:
- I don't really love the the Fractal component browser, but that's just like my opinion
- Less dogfooding of our components, but now that we're using components in many places this is probably less important