ENHANCEMENT: allow component CSS  to proceed
Not really a bug, but something that is obnoxious particuarly in sites that consume vf-core.

Currently if a component's index.scss's dependency isn't present the genenration of the the `vf-component.css` fails, which makes sense -- however it crashes the gulp taks for all the other components, this allows them to still be made.

That is: if `vf-component-1/index.scss` doesn't build, this will allow the build to continue for `vf-component-2/index.scss`