Require vf-sass-config, vf-design-tokens
You can't do anything useful with vf-core without these two components, so it saves a couple of `yarn add`s add removes a common miss-configuration step.

For developing vf-core, this won't mean anything practical, however it does help when using it as a dependency.

Also .npmignore-ing yarn.lock makes vf-core a lighter dependency and ensures that the required vf components are minimums, that is: default to vf-sass-config@v1.0.1, but newer is ok too