COMPONENT - EBI VF 1.x compatibility
This is a port of the CSS we've been using for compatibility. It uses the fixes from vf-wp and requires that a section of html being wrapped with the class `.ebi-vf1-integration` to prevent "accidental fixes".

There's also a little demo page in the NJK file that demonstrates what it does, acts as a nice "is the fix working or still needed" check.

I've also been using these fixes on some EBI pages, so getting this in the mono-repo would be helpful and stuff.