Component/vf tree enhancements (#597)
* Add placeholders for `collpased`

Although it does nothing, I think this makes the behavour easier to understand.

* A bit more of a programmatic naming convention

* Respect the expanded state

* Improve demo-ability of YML

Rather than saying `sublist: true`, add a stub list and set it to not expanded

* Use only expanded

Probably simpler than using `expanded` in some places and `collpased` in others.  Maybe should also change `.vf-tree--collapsed`?

* In place of `currentParent` and `currentChild` use `activeTrail`

* Optimise NJK file

This is probably a pre-mature optimisation, but I think it demonstrates how we can make the code tidier with the standardising of the Yaml values.

* Protodocumentation

Some notes that I hade made.