Bug: Fix Fractal config path
In #613 David noticed there was something off with the Fractal config path.

This PR fixes the behaviour to be as intended: for vf-core to always use the vf-core's `fractal.js` config (client projects aren't expected to bring their own Fractal config).

This fixes that bug, but also raises a couple things:

- Option A: Should `vf-build` be only used by `vf-core`? I think David's experience suggests it has value to other projects
- Obtion B: We support client projects using `vf-build`, but we need to add some paramater so that the correct Fractal build mode is invoked (or not at all, when it's not needed)

That's probably a seperate issue/PR -- best to get this bug fixed first as although `vf-build` could be faster, it at least works with this fix.