the one that makes header components a little better (#836)
* changes component title to simply header

because there's several 'headers' it was confusing that the header container vf-header had a title of site header and not header

* updates header spacings

a mojority of the header components were using grid for centering the content. This replaces that with good old margin: 0 auto and a max-width

* updates header spacings

a mojority of the header components were using grid for centering the content. This replaces that with good old margin: 0 auto and a max-width