Enhanchement: Backwards compatibility with vf-global-header 1.x HTML
Add support for backwards compatibility with .vf-global-header 1.x's vf-global-header__inner

Because the EMBL global header comes from the contentHub we can't switch over everything in one go. This allows projects to update to vf-global-header and the contentHub to continue to have HTML with vf-global-header__inner.

I note that it will be removed in vf-global-header 3.0