chore:  better split out build dependencies
There were still many depencies assigned to vf-core that are actually part of the build process. This is actually not just for Sass build, but JS and Nunjucks compilation. But it will be easier to do future optimsation by moving them to vf-sass-compilation for the time being.

Also managed to spot a couple unsued dependencies.