Enhancment: echo component generation time
This addresses #974 so that we get:

```css
/*! Component: @visual-framework/vf-componenet-rollup
 * Version: 1.2.0
 * Location: components/vf-componenet-rollup
 * VF Core version in use: 2.2.7
 * Build time: Mon, 19 Oct 2020 07:53:11 GMT
 */
```

It adds support in vf-core, the component generator and the vf-component-rollup. To do the reaminder of the components, this line will needto be added to each component's Sass file:

```
 * Build time: #{map-get($componentInfo, 'buildTimeStamp')}
```

Which will of course bump all the versions, perhaps best saved for a less busy time.