enhancement: component library navigation
This continues the ongoing work to improve the component library as discussed in #1175:

- begins to vf-tree ify the navigation
- processes the README.md to present a more readable overview
- moves verbose code readouts and install text to vf-details
- puts code examples in "boxes"
- general tidy

More to be done but I think it's a good stride in the right direction.

Will be follow on work to tidy up the READMEs

Could merge this one now and revisit later or keep stashed and work here.