Refactor: vf-frctl-theme
With the recent changes in the fractal theming layer (see: #1260) our old vf-frctl-theme is no longer sustainable and needs a complete refactor.

The best way to do this is to take the revised frctl-mandlebrot theme and merge in our old code.

This presents and opportunity to reasses vf-frctl-theme to be more vf-core development focused. Especially as the theme is no longer deployed to the web.

High level ideas:

- something closer to the default mandlebrot theme (future sustainability)
- preserve some of the mandlebrot features (searching, resizing)
- apply VF things where it matters (fonts, colours, key principles)

This commit starts that work.