Enhancement: Add vf-news-container--featured
This addresses the long-lingering #1036, where we've been sprinkling local CSS to achieve a "Featured news" container.

Even though the vf-summary and vf-news-container have uncertain futures, this would stop the propogation of "local CSS" across sites to enable a smoother transition to "the new thing", once we figure out what it is.