Docs: create a basic example EBI page
This doesn't do anything exotic, but creates a standard reference of how an EBI page has the black bar and the VF 2.0 footer.

More work to be done but will ease a lot of confusion by EBI devs.

It also updates the _data/containers to link to the njk templates in a way that works for vf-component-library.