the one that refactors lots of form things (#1228)
* style(vf-form__checkbox): resizes checkbox and adds better hover state

We have made the checkbox smaller. Going from 24px to 16px. We have also introduced a better hover
state where the tick is shown on hover

* docs(vf_form__core--example): updates the example page for all vf-form components

Added the vf-stack component to control the spacing

* fix(vf-form__helper): removes the bottom margin on the error helper

This relates to the change in vf-form__item where we are now using vf-stack to control spacing

BREAKING CHANGE: This requires the updated vf-form__item to control the spacing

* style(vf-form__input): refreshes styling of input and required asterick, adds vf-stack to examples

We have reduced the border size and colour of the inputs and reduced the internal spacing. We have
also added the vf-stack element to the examples in the nunjucks file. The reqiured astericks has
also been repositioned to match the new spacing of the input

* feat(vf-form__item): adds vf-stack to the wrapper, changes inline variant, adds custom stack margin

Added the vf-stack component to the wrapper so that we can space the elements of a form item inside
it. Adds a custom stack spacing unit to determine the size of margin. Makes the inline variant tell
the labels to use inline-flex

* style(vf-form__label): removes margins, adds flex and align items

* style(vf-form__radio): makes checked button smaller circle inside border, adds hover state

* style(vf-form__select): reduces border size, changes color, matches padding of vf-form__input

* style(vf-form__textarea): updates border size and colour, and padding to match other inputs

* removes placeholders because idk UX

* fixes error styling and form item hierarchy

* changes hover and focus effect for radio and checkboxes

* fixes issues with inline checkboxes and radios

* all the changelogs

* Typo

* Changelog formatting, github PR link

* Remove placeholder changelog text

* fixes 04px and makes them rems too

* adds hover/focus shadow like the cards

Co-authored-by: Ken Hawkins <khawkins98@gmail.com>