fix: vf-logo extreme
A while back (when the EMBL logo sizing was being tweaked) all the vf-logo variants took on the extreme size. This sorts that.

Idea: As the vf-logo is a generic component that can take any logo, we may also want to add a custom property to allow specifying a size.