Documentations: web content guidelines
As has come up a few times (including #1287) there is a need for some high level guidelines on web content.

For now I've stuck them in the guidance section, but I suspect we have a reorganisation coming sometime soon, and I didn't want to get distracted with that here — instead I've focused on making the content.

Topics added in this PR:

- read more
- icons
- links in new tabs
- white space

This builds atop #1377, which should be merged first.