Update accessibility.md
A few more editorial changes. 
I also moved the sentence about the screen reader under "tools" (and added a link on how to turn on the screen reader) because I think placing it next to "test with real humans" may actually undermine the latter point (i.e. people may think that the screen reader is a sufficient substitute of a person).