enhancement: update colour tokens
* Use revised primary and secondary text colours
* Generate json for theme colours
* Display theme colours in component library
* Deprecates old primary, secondary, tertiary themes in favour of vf-color__brand and interactive, text colours
* Changes mixin `set-type` to use the primary colour (off black) by default

We had also discussed moving to a "functional" name for these colours. For the moment I've left as "theme" as it would require more technical change and I wory it might be confused with "functions" in a programming sense... To be revisitted.

The newer colour tokens still need to mapped into compoentns